+++ short_term/pipeline.py
@@ -1,8 +1,8 @@
 """Whole pipeline script for short-term storage."""
 
-from extract import extract_recordings
-from transform import transform_recordings
-from load import upload_readings
+from short_term.extract_short import extract_recordings
+from short_term.transform_short import transform_recordings
+from short_term.load_short import upload_readings
 
 
 def lambda_handler(event=None, context=None):
